requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
